+++ app/shared/services.js
@@ -43,7 +43,14 @@
 
         // Pass in the user's language selection.
         parameters.user_locale = LanguageService.getLocale();
+
+        // Pass in the account_id, which is required when asking for a token.
         parameters.account_id = settings.account.account_id;
+
+        // If this is a test app, send a test flag to request a test token.
+        if (settings.account.test) {
+            parameters.test = true;
+        }
 
         // Prepare the url
         var endpoint = buildUrl("/auths/limited", settings);
@@ -70,7 +77,15 @@
 
             deferred.resolve(response.data.token);
         }, function (error) {
-            deferred.reject({ type: "internal_server_error", reference: "6lnOOW1", code: "unspecified_error", message: gettextCatalog.getString("There was a problem obtaining authorization for this session. Please reload the page to try your request again."), status: error.status });
+
+            var msg = gettextCatalog.getString("There was a problem obtaining authorization for this session. Please reload the page to try your request again.");
+
+            // If this is a 403 error and you are in test mode, add a note to the error message about test orders.
+            if (settings.account.test && error.data.error.status == 403 && error.data.error.code == "insufficient_permissions") {
+                msg = "This app is installed in test mode and can only be run by authorized test users. To run this app, launch it from within your account while in test mode. If you would like to allow unauthenticated users to run apps in test mode, sign into your account, and enable 'Allow Public Test Orders' under Settings> Technical.";
+            }
+
+            deferred.reject({ type: "internal_server_error", reference: "6lnOOW1", code: "unspecified_error", message: msg, status: error.status });
         });
 
         return deferred.promise;
@@ -802,7 +817,7 @@ app.service("CartService", ['$http', '$q', '$rootScope', 'ApiService', 'PaymentS
 
     }
 
-    function pay(cart, payment_method, parameters, quiet) {
+    function pay(cart, payment_method, parameters, cartParameters, quiet) {
 
         var deferred = $q.defer();
         parameters = setDefaultParameters(parameters);
@@ -828,18 +843,28 @@ app.service("CartService", ['$http', '$q', '$rootScope', 'ApiService', 'PaymentS
             });
         };
 
+        var copyObject = function (cart, newCart) {
+            for (var property in newCart) {
+                if (newCart.hasOwnProperty(property)) {
+                    cart[property] = newCart[property];
+                }
+            }
+        }
+
         // If there currently is no cart, create it. Otherwise, update the existing cart.
-        if (cart.cart_id == null) {
-            create(cart, parameters, quiet).then(function (cart) {
+        if (!cart || cart.cart_id == null) {
+            create(cart, cartParameters, quiet).then(function (data) {
+                copyObject(cart, data);
                 sendPayment(cart.cart_id, payment_method);
             }, function (error) {
                 deferred.reject(error);
             });
-
         } else {
-            update(cart, parameters, quiet).then(function (cart) {
+            update(cart, cartParameters, quiet).then(function (data) {
+                copyObject(cart, data);
                 sendPayment(cart.cart_id, payment_method);
             }, function (error) {
+                copyObject(cart, data);
                 deferred.reject(error);
             });
         }
@@ -1710,6 +1735,10 @@ app.service("LanguageService", ['$q', '$rootScope', 'SettingsService', 'StorageS
         }
 
         StorageService.set("language", language);
+
+        var languages = getLanguages();
+        $rootScope.language = _.find(languages, function (l) { return l.code == language });
+
         gettextCatalog.setCurrentLanguage(language);
 
         // Emit the change
@@ -1888,6 +1917,7 @@ app.service("HelperService", ['SettingsService', 'StorageService', '$location', 
         isRequiredCustomerField: isRequiredCustomerField,
         isOptionalCustomerField: isOptionalCustomerField,
         isCustomerField: isCustomerField,
+        hasRequiredFields: hasRequiredFields,
         hasShippingAddress: hasShippingAddress,
         newSessionRedirect: newSessionRedirect,
         getShoppingUrl: getShoppingUrl,
@@ -1980,6 +2010,24 @@ app.service("HelperService", ['SettingsService', 'StorageService', '$location', 
         }
 
         return false;
+
+    }
+
+    function hasRequiredFields(customer, options) {
+
+        for (i = 0; i < options.customer_required_fields.length; i++) {
+            if (options.customer_required_fields[i].substring(0, 16) == "billing_address.") {
+                if (!customer.billing_address[options.customer_required_fields[i].substring(16)]) {
+                    return false;
+                }
+            } else {
+                if (!customer[options.customer_required_fields[i]]) {
+                    return false;
+                }
+            }
+        }
+
+        return true;
 
     }